+++ OpenGL_Motion_Blur/src/tests/TestClearColor.cpp
@@ -17,7 +17,7 @@ namespace test {
 
 	}
 
-	void TestClearColor::OnUpdate(float deltaTime)
+	void TestClearColor::OnUpdate(double deltaTime)
 	{
 
 	}
