+++ OpenGL_Motion_Blur/src/tests/TestDepthBuffer.h
@@ -35,7 +35,7 @@ namespace test {
 		TestDepthBuffer();
 		~TestDepthBuffer();
 
-		void OnUpdate(float deltaTime) override;
+		void OnUpdate(double deltaTime) override;
 		void OnRender() override;
 		void OnImGuiRender() override;
 	};
